+++ util/logging/messages/generalmsg.go
@@ -192,16 +192,19 @@ type logFieldsMessage struct {
 }
 
 func (l logFieldsMessage) WriteToConsole(console io.Writer) {
-	if l.console || l.serverConsole {
+	if l.console {
 		fmt.Fprintf(console, "%s\n", l.composeMsg())
-		if l.serverConsole {
-			if l.jsonObj != nil {
-				fmt.Fprintf(console, "%s\n", l.jsonObj.StringIndent("", " "))
-			}
+	}
+}
 
-			fmt.Fprintf(console, "Fields: %s\n", strings.Join(l.Fields, ","))
-			fmt.Fprintf(console, "Types: %s\n", strings.Join(l.Types, ","))
+func (l logFieldsMessage) WriteExtraToConsole(console io.Writer) {
+	if l.console {
+		if l.jsonObj != nil {
+			fmt.Fprintf(console, "%s\n", l.jsonObj.StringIndent("", " "))
 		}
+
+		fmt.Fprintf(console, "Fields: %s\n", strings.Join(l.Fields, ","))
+		fmt.Fprintf(console, "Types: %s\n", strings.Join(l.Types, ","))
 	}
 }
 
@@ -231,7 +234,7 @@ func (l logFieldsMessage) returnIDs() (ids string, values string) {
 func (l logFieldsMessage) composeMsg() string {
 	return l.msg
 }
-func buildLogFieldsMessage(msg string, console bool, serverConsole bool, withIDs bool, loglvl logging.Level) logFieldsMessage {
+func buildLogFieldsMessage(msg string, console bool, withIDs bool, loglvl logging.Level) logFieldsMessage {
 	logMessage := logFieldsMessage{
 		CallerInfo:         logging.GetCallerInfo(logging.NotHere()),
 		Level:              loglvl,
@@ -240,7 +243,6 @@ func buildLogFieldsMessage(msg string, console bool, serverConsole bool, withIDs
 		JSONRepresentation: "",
 		msg:                msg,
 		console:            console,
-		serverConsole:      serverConsole,
 		withIDs:            withIDs,
 	}
 
@@ -255,7 +257,7 @@ func buildLogFieldsMessage(msg string, console bool, serverConsole bool, withIDs
 
 //ReportLogFieldsMessageWithIDs report message with Ids
 func ReportLogFieldsMessageWithIDs(msg string, loglvl logging.Level, logSink logging.LogSink, items ...interface{}) {
-	logMessage := buildLogFieldsMessage(msg, false, false, true, loglvl)
+	logMessage := buildLogFieldsMessage(msg, false, true, loglvl)
 	logMessage.CallerInfo.ExcludeMe()
 	logMessage.reportLogFieldsMessage(logSink, items...)
 
@@ -263,21 +265,14 @@ func ReportLogFieldsMessageWithIDs(msg string, loglvl logging.Level, logSink log
 
 //ReportLogFieldsMessageToConsole report message to console
 func ReportLogFieldsMessageToConsole(msg string, loglvl logging.Level, logSink logging.LogSink, items ...interface{}) {
-	logMessage := buildLogFieldsMessage(msg, true, false, false, loglvl)
-	logMessage.CallerInfo.ExcludeMe()
-	logMessage.reportLogFieldsMessage(logSink, items...)
-}
-
-//ReportLogFieldsMessageToServerConsole report message so that console will write out json and ids/fields
-func ReportLogFieldsMessageToServerConsole(msg string, loglvl logging.Level, logSink logging.LogSink, items ...interface{}) {
-	logMessage := buildLogFieldsMessage(msg, true, true, false, loglvl)
+	logMessage := buildLogFieldsMessage(msg, true, false, loglvl)
 	logMessage.CallerInfo.ExcludeMe()
 	logMessage.reportLogFieldsMessage(logSink, items...)
 }
 
 //ReportLogFieldsMessage generate a logFieldsMessage log entry
 func ReportLogFieldsMessage(msg string, loglvl logging.Level, logSink logging.LogSink, items ...interface{}) {
-	logMessage := buildLogFieldsMessage(msg, false, false, true, loglvl)
+	logMessage := buildLogFieldsMessage(msg, false, true, loglvl)
 	logMessage.CallerInfo.ExcludeMe()
 
 	logMessage.reportLogFieldsMessage(logSink, items...)