+++ util/logging/messages/generalmsg_test.go
@@ -19,9 +19,9 @@ func TestReportLogFieldsMessage_Complete(t *testing.T) {
 		},
 		logging.StandardVariableFields,
 		map[string]interface{}{
-			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server",
+			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,ExtraConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server",
 			"sous-types":     "Config,string,bool",
-			"json-value":     "{\"message\":{\"array\":[\"(logging.Config) {\\n Basic: (struct { Level string \\\"env:\\\\\\\"SOUS_LOGGING_LEVEL\\\\\\\"\\\"; DisableConsole bool }) {\\n  Level: (string) \\\"\\\",\\n  DisableConsole: (bool) false\\n },\\n Kafka: (struct { Enabled bool; DefaultLevel string \\\"env:\\\\\\\"SOUS_KAFKA_LOG_LEVEL\\\\\\\"\\\"; Topic string \\\"env:\\\\\\\"SOUS_KAFKA_TOPIC\\\\\\\"\\\"; Brokers []string; BrokerList string \\\"env:\\\\\\\"SOUS_KAFKA_BROKERS\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  DefaultLevel: (string) \\\"\\\",\\n  Topic: (string) (len=10) \\\"test-topic\\\",\\n  Brokers: ([]string) \\u003cnil\\u003e,\\n  BrokerList: (string) (len=23) \\\"broker1,broker2,broker3\\\"\\n },\\n Graphite: (struct { Enabled bool; Server string \\\"env:\\\\\\\"SOUS_GRAPHITE_SERVER\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  Server: (string) \\\"\\\"\\n }\\n}\\n\"]}}",
+			"json-value":     "{\"message\":{\"array\":[\"(logging.Config) {\\n Basic: (struct { Level string \\\"env:\\\\\\\"SOUS_LOGGING_LEVEL\\\\\\\"\\\"; DisableConsole bool; ExtraConsole bool \\\"env:\\\\\\\"SOUS_EXTRA_CONSOLE\\\\\\\"\\\" }) {\\n  Level: (string) \\\"\\\",\\n  DisableConsole: (bool) false,\\n  ExtraConsole: (bool) false\\n },\\n Kafka: (struct { Enabled bool; DefaultLevel string \\\"env:\\\\\\\"SOUS_KAFKA_LOG_LEVEL\\\\\\\"\\\"; Topic string \\\"env:\\\\\\\"SOUS_KAFKA_TOPIC\\\\\\\"\\\"; Brokers []string; BrokerList string \\\"env:\\\\\\\"SOUS_KAFKA_BROKERS\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  DefaultLevel: (string) \\\"\\\",\\n  Topic: (string) (len=10) \\\"test-topic\\\",\\n  Brokers: ([]string) \\u003cnil\\u003e,\\n  BrokerList: (string) (len=23) \\\"broker1,broker2,broker3\\\"\\n },\\n Graphite: (struct { Enabled bool; Server string \\\"env:\\\\\\\"SOUS_GRAPHITE_SERVER\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  Server: (string) \\\"\\\"\\n }\\n}\\n\"]}}",
 			"@loglov3-otl":   "sous-generic-v1",
 			"sous-ids":       "",
 			"sous-id-values": "",
@@ -70,9 +70,9 @@ func TestReportLogFieldsMessage_StructAndString(t *testing.T) {
 		},
 		logging.StandardVariableFields,
 		map[string]interface{}{
-			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server",
 			"sous-types":     "Config,string,bool",
-			"json-value":     "{\"message\":{\"array\":[\"(logging.Config) {\\n Basic: (struct { Level string \\\"env:\\\\\\\"SOUS_LOGGING_LEVEL\\\\\\\"\\\"; DisableConsole bool }) {\\n  Level: (string) \\\"\\\",\\n  DisableConsole: (bool) false\\n },\\n Kafka: (struct { Enabled bool; DefaultLevel string \\\"env:\\\\\\\"SOUS_KAFKA_LOG_LEVEL\\\\\\\"\\\"; Topic string \\\"env:\\\\\\\"SOUS_KAFKA_TOPIC\\\\\\\"\\\"; Brokers []string; BrokerList string \\\"env:\\\\\\\"SOUS_KAFKA_BROKERS\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  DefaultLevel: (string) \\\"\\\",\\n  Topic: (string) (len=10) \\\"test-topic\\\",\\n  Brokers: ([]string) \\u003cnil\\u003e,\\n  BrokerList: (string) (len=23) \\\"broker1,broker2,broker3\\\"\\n },\\n Graphite: (struct { Enabled bool; Server string \\\"env:\\\\\\\"SOUS_GRAPHITE_SERVER\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  Server: (string) \\\"\\\"\\n }\\n}\\n\",\"{\\\"string\\\":{\\\"string\\\":\\\"simple string\\\"}}\"]}}",
+			"json-value":     "{\"message\":{\"array\":[\"(logging.Config) {\\n Basic: (struct { Level string \\\"env:\\\\\\\"SOUS_LOGGING_LEVEL\\\\\\\"\\\"; DisableConsole bool; ExtraConsole bool \\\"env:\\\\\\\"SOUS_EXTRA_CONSOLE\\\\\\\"\\\" }) {\\n  Level: (string) \\\"\\\",\\n  DisableConsole: (bool) false,\\n  ExtraConsole: (bool) false\\n },\\n Kafka: (struct { Enabled bool; DefaultLevel string \\\"env:\\\\\\\"SOUS_KAFKA_LOG_LEVEL\\\\\\\"\\\"; Topic string \\\"env:\\\\\\\"SOUS_KAFKA_TOPIC\\\\\\\"\\\"; Brokers []string; BrokerList string \\\"env:\\\\\\\"SOUS_KAFKA_BROKERS\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  DefaultLevel: (string) \\\"\\\",\\n  Topic: (string) (len=10) \\\"test-topic\\\",\\n  Brokers: ([]string) \\u003cnil\\u003e,\\n  BrokerList: (string) (len=23) \\\"broker1,broker2,broker3\\\"\\n },\\n Graphite: (struct { Enabled bool; Server string \\\"env:\\\\\\\"SOUS_GRAPHITE_SERVER\\\\\\\"\\\" }) {\\n  Enabled: (bool) false,\\n  Server: (string) \\\"\\\"\\n }\\n}\\n\",\"{\\\"string\\\":{\\\"string\\\":\\\"simple string\\\"}}\"]}}",
+			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,ExtraConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server",
 			"@loglov3-otl":   "sous-generic-v1",
 			"sous-ids":       "",
 			"sous-id-values": "",
@@ -93,8 +93,8 @@ func TestReportLogFieldsMessage_TwoStructs(t *testing.T) {
 		},
 		append(logging.StandardVariableFields, "json-value"),
 		map[string]interface{}{
-			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server,Status,StatusCode,Proto,ProtoMajor,ProtoMinor,Header,Body,ContentLength,TransferEncoding,Close,Uncompressed,Trailer,Request,TLS,Response", //nolint
 			"sous-types":     "Config,string,bool,*Response,int,Header,int64,*Request,*ConnectionState",
+			"sous-fields":    "Basic,Kafka,Graphite,Config,Level,DisableConsole,ExtraConsole,Enabled,DefaultLevel,Topic,Brokers,BrokerList,Server,Status,StatusCode,Proto,ProtoMajor,ProtoMinor,Header,Body,ContentLength,TransferEncoding,Close,Uncompressed,Trailer,Request,TLS,Response",
 			"@loglov3-otl":   "sous-generic-v1",
 			"sous-ids":       "",
 			"sous-id-values": "",
@@ -181,7 +181,7 @@ type TestID struct {
 }
 
 func TestExtractIDs_TwoIds(t *testing.T) {
-	l := buildLogFieldsMessage("this is a test", false, false, true, logging.ExtraDebug1Level)
+	l := buildLogFieldsMessage("this is a test", false, true, logging.ExtraDebug1Level)
 
 	d := &TestID{
 		TestInnerID: TestInnerID{
@@ -198,7 +198,7 @@ func TestExtractIDs_TwoIds(t *testing.T) {
 }
 
 func TestExtractIDs_NoIds(t *testing.T) {
-	l := buildLogFieldsMessage("this is a test", false, false, true, logging.ExtraDebug1Level)
+	l := buildLogFieldsMessage("this is a test", false, true, logging.ExtraDebug1Level)
 
 	foo := "hello"
 	l.extractID(foo)