+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface.
 
+
+## [Unreleased](//github.com/opentable/sous/compare/0.5.76...HEAD)
+
+### Fixed
+* Server: /deploy-queue-item now returns correct queue position and Resolution field.
+
 ## [0.5.76](//github.com/opentable/sous/compare/0.5.72...0.5.76)
 
 ### Added
@@ -16,7 +22,7 @@ with respect to its command line interface and HTTP interface.
 * Server now accepts -autoresolver=false to disable to autoresolver.
 * All: Added ability to add context to http requests
 * Bugfixes for the newdeploy command
-	
+
 ### Fixed
 * Client: running `sous` from outside of git workspaces no longer results in
   a confusing Git error.